added a top reaction to QuestRecommendationMessage, moved QuestAlgorithm to SpecificQuestAlgorithm
