+++ src/main/java/apple/questing/discord/commands/CommandQuestSpecific.java
@@ -1,7 +1,5 @@
 package apple.questing.discord.commands;
 
-import apple.questing.QuestAlgorithm;
-import apple.questing.data.FinalQuestOptions;
 import apple.questing.data.WynncraftClass;
 import apple.questing.data.WynncraftPlayer;
 import apple.questing.data.reaction.AllReactableClassChoices;
