+++ src/main/java/apple/questing/sheets/SheetsQuery.java
@@ -1,6 +1,8 @@
 package apple.questing.sheets;
 
+import apple.questing.QuestMain;
 import apple.questing.data.Quest;
+import com.google.api.services.sheets.v4.model.Spreadsheet;
 import com.google.api.services.sheets.v4.model.ValueRange;
 
 import java.io.IOException;
